streaming_etl/requirements.txt
pandas
numpy
pyyaml
python-dotenv
pyfiglet
